Processing pipeline
The ICr processing pipeline script was written to take a folder containing one or more datasets and process them from integration to refinement.

The script will run indexing/integration in DIALS, scaling/merging with AIMLESS, molecular replacement with PHASER and refinement with BUSTER.

An input file should be filled with the appropriate details.
Example usage:
./ICR_processing_pipeline.sh ./ICR_pipeline.INP