EDSTATS
Scripts to be run on a dose series refined with Buster to produce a sequence of refined models and reflection files that can be used to generate 2Fo-Fc and Fo-Fc maps. These maps are input to EDSTATS to allow the local fit of the model to the electron desnity to be monitored as a function of X-ray dose.